requirements.txt
aiogram~=3.13.1
sqlalchemy~=2.0.35
asyncpg
python-dotenv~=1.0.1
psycopg2-binary
playwright